requirements.txt
# Core data pipeline dependencies
dlt[duckdb,parquet]>=0.4.0
requests>=2.31.0
PyYAML>=6.0
pandas>=2.1.0
pyarrow>=14.0.0

# Optional: for future Silver layer and AI agent
# duckdb>=0.9.0
# llama-index>=0.9.0

# Development and testing
pytest>=7.4.0
python-dotenv>=1.0.0
